Network: (somewhat) robust if connections drop out
Keep track of in-flight work units and reschedule them on other nodes if
the node that is working on them drops out or is taking too long to
respond to a ping request.

Currently still has trouble recovering from some specific connection
issues (e.g. if the very last worker that had active work drops out when
the other workers have already completed, then we don't wake a worker
that was 'ready' to retry the failed work unit -- yet.)